refactor: expose service-icon utility
