+++ libs/domains/services/feature/src/index.ts
@@ -49,4 +49,5 @@ export * from './lib/dockerfile-settings/dockerfile-settings'
 export * from './lib/confirmation-cancel-lifecycle-modal/confirmation-cancel-lifecycle-modal'
 export * from './lib/service-template-indicator/service-template-indicator'
 export * from './lib/service-avatar/service-avatar'
+export * from './lib/service-icon/service-icon'
 export * from './lib/service-access-modal/service-access-modal'
